user related changes; shared system support

+++ src/main/java/com/dimazombie/famtree/model/JDBCNodeRepository.java
@@ -6,6 +6,8 @@ import org.slf4j.LoggerFactory;
 import javax.naming.InitialContext;
 import javax.naming.NamingException;
 import javax.sql.DataSource;
+import javax.ws.rs.container.ContainerRequestContext;
+import javax.ws.rs.core.Context;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
@@ -13,7 +15,10 @@ import java.util.List;
 public class JDBCNodeRepository implements NodeRepository {
     private Logger logger = LoggerFactory.getLogger(NodeRepository.class);
 
+    @Context ContainerRequestContext context;
+
     public List<Node> getAllNodes() {
+        User user = (User) context.getProperty(User.PROPERTY_NAME);
         Connection conn = null;
         try {
             List<Node> nodes = new ArrayList<Node>();
@@ -22,10 +27,14 @@ public class JDBCNodeRepository implements NodeRepository {
             DataSource ds = (DataSource) ctx.lookup("jdbc/ds");
             conn = ds.getConnection();
 
-            String sql = "select t.id, t.parent_id, t.name, t.bio, t.image_id, t.date_of_birth from NODE t where t.parent_id is null";
+            String sql = "select t.id, t.parent_id, t.name, t.bio, t.image_id, t.date_of_birth " +
+                    "from NODE t " +
+                    "where t.parent_id is null and t.user_id = ?";
             logger.debug("SQL: " + sql);
+            logger.debug("with bind var: " + user.getId());
 
             PreparedStatement stmt = conn.prepareStatement(sql);
+            stmt.setLong(1, user.getId());
 
             ResultSet rs = stmt.executeQuery();
             while(rs.next()) {
@@ -64,6 +73,7 @@ public class JDBCNodeRepository implements NodeRepository {
 
     @Override
     public Node findById(String nodeId) {
+        User user = (User) context.getProperty(User.PROPERTY_NAME);
         Connection conn = null;
         try {
             Node node = null;
@@ -72,12 +82,16 @@ public class JDBCNodeRepository implements NodeRepository {
             DataSource ds = (DataSource) ctx.lookup("jdbc/ds");
             conn = ds.getConnection();
 
-            String sql = "select t.id, t.parent_id, t.name, t.bio, t.image_id, t.date_of_birth from NODE t where t.id = ?";
+            String sql = "select t.id, t.parent_id, t.name, t.bio, t.image_id, t.date_of_birth " +
+                    "from NODE t " +
+                    "where t.id = ? and t.user_id = ?";
             logger.debug("SQL: " + sql);
             logger.debug("with bind var: " + nodeId);
+            logger.debug("with bind var: " + user.getId());
 
             PreparedStatement stmt = conn.prepareStatement(sql);
             stmt.setString(1, nodeId);
+            stmt.setLong(2, user.getId());
 
             ResultSet rs = stmt.executeQuery();
             if(rs.next()) {
@@ -107,14 +121,19 @@ public class JDBCNodeRepository implements NodeRepository {
     }
 
     private List<Node> getAncestors(Long nodeId, Connection conn) throws SQLException {
+        User user = (User) context.getProperty(User.PROPERTY_NAME);
         List<Node> nodes = new ArrayList<Node>();
 
-        String sql = "select t.id, t.parent_id, t.name, t.bio, t.image_id, t.date_of_birth from NODE t where t.parent_id = ?";
+        String sql = "select t.id, t.parent_id, t.name, t.bio, t.image_id, t.date_of_birth " +
+                "from NODE t " +
+                "where t.parent_id = ? and t.user_id = ?";
         logger.debug("SQL: " + sql);
         logger.debug("with bind var: " + nodeId);
+        logger.debug("with bind var: " + user.getId());
 
         PreparedStatement stmt = conn.prepareStatement(sql);
         stmt.setLong(1, nodeId);
+        stmt.setLong(2, user.getId());
 
         ResultSet rs = stmt.executeQuery();
         while (rs.next()) {
@@ -145,6 +164,7 @@ public class JDBCNodeRepository implements NodeRepository {
     }
 
     public Node persist(Node node) {
+        User user = (User) context.getProperty(User.PROPERTY_NAME);
         Connection conn = null;
         try {
             InitialContext ctx = new InitialContext();
@@ -152,13 +172,14 @@ public class JDBCNodeRepository implements NodeRepository {
             conn = ds.getConnection();
 
             if(node.getId() == null) {
-                String sql = "insert into NODE(parent_id, name, bio, image_id, date_of_birth) values (?, ?, ?, ?, ?)";
+                String sql = "insert into NODE(parent_id, name, bio, image_id, date_of_birth, user_id) values (?, ?, ?, ?, ?, ?)";
                 logger.debug("SQL: " + sql);
                 logger.debug("with bind var: " + node.getParentId());
                 logger.debug("with bind var: " + node.getName());
                 logger.debug("with bind var: " + node.getBio());
                 logger.debug("with bind var: " + node.getImageId());
                 logger.debug("with bind var: " + node.getDateOfBirth());
+                logger.debug("with bind var: " + user.getId());
 
                 PreparedStatement stmt = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
                 if (node.getParentId() != null) {
@@ -174,6 +195,7 @@ public class JDBCNodeRepository implements NodeRepository {
                     stmt.setNull(4, Types.BIGINT);
                 }
                 stmt.setString(5, node.getDateOfBirth());
+                stmt.setLong(6, user.getId());
                 stmt.execute();
 
                 ResultSet rs = stmt.getGeneratedKeys();
@@ -184,7 +206,7 @@ public class JDBCNodeRepository implements NodeRepository {
                 logger.debug("rs: " + node);
             } else {
                 String sql = "update NODE t set t.parent_id = ?, t.name = ?, t.bio = ?, t.image_id = ?, " +
-                        "t.date_of_birth = ? where t.id = ?";
+                        "t.date_of_birth = ? where t.id = ? and t.user_id = ?";
                 logger.debug("SQL: " + sql);
                 logger.debug("with bind var: " + node.getParentId());
                 logger.debug("with bind var: " + node.getName());
@@ -192,6 +214,7 @@ public class JDBCNodeRepository implements NodeRepository {
                 logger.debug("with bind var: " + node.getImageId());
                 logger.debug("with bind var: " + node.getDateOfBirth());
                 logger.debug("with bind var: " + node.getId());
+                logger.debug("with bind var: " + user.getId());
 
                 PreparedStatement stmt = conn.prepareStatement(sql);
                 if (node.getParentId() != null) {
@@ -208,6 +231,7 @@ public class JDBCNodeRepository implements NodeRepository {
                 }
                 stmt.setString(5, node.getDateOfBirth());
                 stmt.setLong(6, node.getId());
+                stmt.setLong(7, user.getId());
                 stmt.execute();
             }
 
@@ -223,18 +247,21 @@ public class JDBCNodeRepository implements NodeRepository {
     }
 
     public void remove(Node node) {
+        User user = (User) context.getProperty(User.PROPERTY_NAME);
         Connection conn = null;
         try {
             InitialContext ctx = new InitialContext();
             DataSource ds = (DataSource) ctx.lookup("jdbc/ds");
             conn = ds.getConnection();
 
-            String sql = "delete from NODE t where t.id = ?";
+            String sql = "delete from NODE t where t.id = ? and t.user_id = ?";
             logger.debug("SQL: " + sql);
             logger.debug("with bind var: " + node.getId());
+            logger.debug("with bind var: " + user.getId());
 
             PreparedStatement stmt = conn.prepareStatement(sql);
             stmt.setLong(1, node.getId());
+            stmt.setLong(2, user.getId());
             stmt.execute();
         } catch (SQLException e) {
             throw new RuntimeException(e);
